6-size.c

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int b;
 long int c;
 long long int d;
 float f;
-printf:("Size of a char: %d bytes(s)\n", (unsigned long)sizeof(a));
-printf("Size of an int: %d bytes(s)\n", (unsigned long)sizeof(b));
-printf("Size of a long int: %d bytes(s)\n", (unsigned long)sizeof(c));
-printf("Size of a long long int: %d bytes(s)\n", (unsigned long)sizeof(d));
-printf("Size of a float: %d bytes(s)\n", (unsigned long)sizeof(f));
+printf:("Size of a char: %d bytes(s)\n", sizeof(a));
+printf("Size of an int: %d bytes(s)\n", sizeof(b));
+printf("Size of a long int: %d bytes(s)\n", sizeof(c));
+printf("Size of a long long int: %d bytes(s)\n", sizeof(d));
+printf("Size of a float: %d bytes(s)\n", sizeof(f));
 return (0);
 }
